Fix negative voltage when Wifi module is off
When the WiFi module was turned off, the power controller did not have
a common ground with the wifi module, which would cause the relative
voltage (from the wifi module's perspective) to be negative. This would
strangely result in a constant low voltage of around 1-1.5 volts being
present at the VCC and GND pins of the WiFi module. This resulted in
weird behavior and overall needlessly higher power consumption.

The fix involves the PB0 pin being set to INPUT so to avoid this noise
on the Wifi module's side. Whenever we want to transmit a pulse through
the PB0 pin, we set it as OUTPUT (then the WiFi module is ON so there is
no problem) and when we are done we set it back to INPUT again.

Fixes #65